UI updates and overall site improvements
added: more cleaner format of site, gridded products with hover animated enlargement, fixed the navigation bar so the hover only covers the text, changed the page splitter format so it fits as the resolution of the browser, favicon to add branding to the tab. animated enlargement to the logo and buttons

todo: product catalog images in a grid element, reviews in a grid element, add footer with business information, create 'back to top' button which scrolls the user back to the top.