Gridded images and created button
added: gridded images at the top of the page with a custom button (rather than using a png button img). Redirected the button to a product catalog placeholder., added hover animations with a custom container holding all of the images and buttons together

todo: grid the reviews and add hover animations, add footer and brand information. add back to the top button