Changed fonts and minor updates
Changed font from a built in font to a google font (referenced in the head) and applied all across the site in various forms, minor changes to formatting such as the text in the 'our collections' grid (this fixes the hover animation bug i had).

todo: add footer with brand information, make all images and the logo reference a link (or change the cursor upon hover), add 'back to top' button near footer (as referenced in the figma designs).

After todo: create a product view page which showcases a product alongside sizes, dynamic images that change based on clicks, product description and 'add to cart' functionality